Ldap.Asn1.Type

Documentation

data LdapMessage op

Message envelope. (Section 4.1.1.)

Constructors

LdapMessage

Fields

ldapMessageId :: !Id

ldapMessageOp :: !op

ldapMessageControls :: !(Maybe Controls)

newtype Id

Every message being processed has a unique non-zero integer ID. (Section 4.1.1.1.)

Constructors

Id

Fields

unId :: Int32

data ProtocolClientOp

Client requests. The RFC doesn't make a difference between ProtocolClientOp and ProtocolServerOp but it's useful to distinguish between them in Haskell.

Constructors

BindRequest !Int8 !LdapDn !AuthenticationChoice

UnbindRequest

SearchRequest !LdapDn !Scope !DerefAliases !Int32 !Int32 !Bool !Filter !AttributeSelection

ModifyRequest !LdapDn ![(Operation, PartialAttribute)]

AddRequest !LdapDn !AttributeList

DeleteRequest !LdapDn

ModifyDnRequest !LdapDn !RelativeLdapDn !Bool !(Maybe LdapDn)

CompareRequest !LdapDn !AttributeValueAssertion

ExtendedRequest !LdapOid !(Maybe ByteString)

data ProtocolServerOp

Server responses. The RFC doesn't make a difference between ProtocolClientOp and ProtocolServerOp but it's useful to distinguish between them in Haskell.

Constructors

BindResponse !LdapResult !(Maybe ByteString)

SearchResultEntry !LdapDn !PartialAttributeList

SearchResultReference !(NonEmpty Uri)

SearchResultDone !LdapResult

ModifyResponse !LdapResult

AddResponse !LdapResult

DeleteResponse !LdapResult

ModifyDnResponse !LdapResult

CompareResponse !LdapResult

ExtendedResponse !LdapResult !(Maybe LdapOid) !(Maybe ByteString)

IntermediateResponse !(Maybe LdapOid) !(Maybe ByteString)

newtype AuthenticationChoice

Not really a choice until SASL is supported.

Constructors

Simple ByteString

data Scope

Scope of the search to be performed.

Constructors

BaseObject
Constrained to the entry named by baseObject.

SingleLevel
Constrained to the immediate subordinates of the entry named by baseObject.

WholeSubtree
Constrained to the entry named by baseObject and to all its subordinates.

data DerefAliases

An indicator as to whether or not alias entries (as defined in [RFC4512]) are to be dereferenced during stages of the Search operation.

Constructors

NeverDerefAliases
Do not dereference aliases in searching or in locating the base object of the Search.

DerefInSearching
While searching subordinates of the base object, dereference any alias within the search scope.

DerefFindingBaseObject
Dereference aliases in locating the base object of the Search.

DerefAlways
Dereference aliases both in searching and in locating the base object of the Search.

data Filter

Conditions that must be fulfilled in order for the Search to match a given entry.

Constructors

And !(NonEmpty Filter)
All filters evaluate to TRUE

Or !(NonEmpty Filter)
Any filter evaluates to TRUE

Not Filter
Filter evaluates to FALSE

EqualityMatch AttributeValueAssertion
EQUALITY rule returns TRUE

Substrings SubstringFilter
SUBSTR rule returns TRUE

GreaterOrEqual AttributeValueAssertion
ORDERING rule returns FALSE

LessOrEqual AttributeValueAssertion
ORDERING or EQUALITY rule returns TRUE

Present AttributeDescription
Attribute is present in the entry

ApproxMatch AttributeValueAssertion
Same as EqualityMatch for most servers

ExtensibleMatch MatchingRuleAssertion

data SubstringFilter

Constructors

SubstringFilter !AttributeDescription !(NonEmpty Substring)

data Substring

Constructors

Initial !AssertionValue

Any !AssertionValue

Final !AssertionValue

data MatchingRuleAssertion

Constructors

MatchingRuleAssertion !(Maybe MatchingRuleId) !(Maybe AttributeDescription) !AssertionValue !Bool

newtype MatchingRuleId

Matching rules are defined in Section 4.1.3 of [RFC4512]. A matching rule is identified in the protocol by the printable representation of either its numericoid or one of its short name descriptors [RFC4512], e.g., caseIgnoreMatch or '2.5.13.2'. (Section 4.1.8.)

Constructors

MatchingRuleId LdapString

newtype AttributeSelection

Constructors

AttributeSelection [LdapString]

newtype AttributeList

Constructors

AttributeList [Attribute]

newtype PartialAttributeList

Constructors

PartialAttributeList [PartialAttribute]

newtype Controls

Constructors

Controls [Control]

data Control

Constructors

Control !LdapOid !Bool !(Maybe ByteString)

data LdapResult

Constructors

LdapResult !ResultCode !LdapDn !LdapString !(Maybe ReferralUris)

data ResultCode

LDAP operation's result.

Constructors

Success

OperationError

ProtocolError

TimeLimitExceeded

SizeLimitExceeded

CompareFalse

CompareTrue

AuthMethodNotSupported

StrongerAuthRequired

Referral

AdminLimitExceeded

UnavailableCriticalExtension

ConfidentialityRequired

SaslBindInProgress

NoSuchAttribute

UndefinedAttributeType

InappropriateMatching

ConstraintViolation

AttributeOrValueExists

InvalidAttributeSyntax

NoSuchObject

AliasProblem

InvalidDNSyntax

AliasDereferencingProblem

InappropriateAuthentication

InvalidCredentials

InsufficientAccessRights

Busy

Unavailable

UnwillingToPerform

LoopDetect

NamingViolation

ObjectClassViolation

NotAllowedOnNonLeaf

NotAllowedOnRDN

EntryAlreadyExists

ObjectClassModsProhibited

AffectsMultipleDSAs

Other

newtype AttributeDescription

Constructors

AttributeDescription LdapString

newtype AttributeValue

Constructors

AttributeValue ByteString

data AttributeValueAssertion

Constructors

AttributeValueAssertion !AttributeDescription !AssertionValue

newtype AssertionValue

Constructors

AssertionValue ByteString

data Attribute

Constructors

Attribute !AttributeDescription !(NonEmpty AttributeValue)

data PartialAttribute

Constructors

PartialAttribute !AttributeDescription ![AttributeValue]

newtype LdapDn

An LDAPDN is defined to be the representation of a Distinguished Name (DN) after encoding according to the specification in [RFC4514].

Constructors

LdapDn LdapString

newtype RelativeLdapDn

A RelativeLDAPDN is defined to be the representation of a Relative Distinguished Name (RDN) after encoding according to the specification in [RFC4514].

Constructors

RelativeLdapDn LdapString

newtype ReferralUris

Constructors

ReferralUris (NonEmpty Uri)

newtype Uri

Constructors

Uri LdapString

data Operation

Constructors

Add

Delete

Replace

newtype LdapString

The LDAPString is a notational convenience to indicate that, although strings of LDAPString type encode as ASN.1 OCTET STRING types, the [ISO10646] character set (a superset of [Unicode]) is used, encoded following the UTF-8 [RFC3629] algorithm. (Section 4.1.2.)

Constructors

LdapString Text

newtype LdapOid

The LDAPOID is a notational convenience to indicate that the permitted value of this string is a (UTF-8 encoded) dotted-decimal representation of an OBJECT IDENTIFIER. Although an LDAPOID is encoded as an OCTET STRING, values are limited to the definition of <numericoid> given in Section 1.4 of [RFC4512].

Constructors

LdapOid Text