Ldap.Client

Description

Pure Haskell LDAP client library.

Documentation

with :: Host -> PortNumber -> (Ldap -> IO a) -> IO (Either LdapError a)

The entrypoint into LDAP.

It catches all LDAP-related exceptions.

data Host

LDAP host.

Constructors

Plain String
Plain LDAP. Do not use!

Insecure String
LDAP over TLS without the certificate validity check. Only use for testing!

Secure String
LDAP over TLS. Use!

data PortNumber :: *

data Ldap

A token. All functions that interact with the Directory require one.

data LdapError

Various failures that can happen when working with LDAP.

Constructors

IOError IOError
Network failure.

ParseError ASN1Error
Invalid ASN.1 data received from the server.

ResponseError ResponseError
An LDAP operation failed.

DisconnectError Disconnect
Notice of Disconnection has been received.

data ResponseError

Response indicates a failed operation.

Constructors

ResponseInvalid Request Response
LDAP server did not follow the protocol, so ldap-client couldn't make sense of the response.

ResponseErrorCode Request ResultCode Dn Text
The response contains a result code indicating failure and an error message.

data ResultCode

LDAP operation's result.

Constructors

Success

OperationError

ProtocolError

TimeLimitExceeded

SizeLimitExceeded

CompareFalse

CompareTrue

AuthMethodNotSupported

StrongerAuthRequired

Referral

AdminLimitExceeded

UnavailableCriticalExtension

ConfidentialityRequired

SaslBindInProgress

NoSuchAttribute

UndefinedAttributeType

InappropriateMatching

ConstraintViolation

AttributeOrValueExists

InvalidAttributeSyntax

NoSuchObject

AliasProblem

InvalidDNSyntax

AliasDereferencingProblem

InappropriateAuthentication

InvalidCredentials

InsufficientAccessRights

Busy

Unavailable

UnwillingToPerform

LoopDetect

NamingViolation

ObjectClassViolation

NotAllowedOnNonLeaf

NotAllowedOnRDN

EntryAlreadyExists

ObjectClassModsProhibited

AffectsMultipleDSAs

Other

data Async a

Asynchronous LDAP operation. Use wait or waitSTM to wait for its completion.

Bind

newtype Password

User's password.

Constructors

Password ByteString

bind :: Ldap -> Dn -> Password -> IO ()

Perform the Bind operation synchronously. Raises ResponseError on failures.

Search

search :: Ldap -> Dn -> Mod Search -> Filter -> [Attr] -> IO [SearchEntry]

Perform the Search operation synchronously. Raises ResponseError on failures.

data SearchEntry

Entry found during the Search.

Constructors

SearchEntry !Dn !(AttrList [])

Search modifiers

data Search

Search options. Use Mod to change some of those.

data Mod a

Search modifier. Combine using Semigroup and/or Monoid instance.

data Scope

Scope of the search to be performed.

Constructors

BaseObject
Constrained to the entry named by baseObject.

SingleLevel
Constrained to the immediate subordinates of the entry named by baseObject.

WholeSubtree
Constrained to the entry named by baseObject and to all its subordinates.

scope :: Scope -> Mod Search

Scope of the search (default: WholeSubtree).

size :: Int32 -> Mod Search

Maximum number of entries to be returned as a result of the Search. No limit if the value is 0 (default: 0).

time :: Int32 -> Mod Search

Maximum time (in seconds) allowed for the Search. No limit if the value is 0 (default: 0).

typesOnly :: Bool -> Mod Search

Whether Search results are to contain just attribute descriptions, or both attribute descriptions and values (default: False).

data DerefAliases

An indicator as to whether or not alias entries (as defined in [RFC4512]) are to be dereferenced during stages of the Search operation.

Constructors

NeverDerefAliases
Do not dereference aliases in searching or in locating the base object of the Search.

DerefInSearching
While searching subordinates of the base object, dereference any alias within the search scope.

DerefFindingBaseObject
Dereference aliases in locating the base object of the Search.

DerefAlways
Dereference aliases both in searching and in locating the base object of the Search.

derefAliases :: DerefAliases -> Mod Search

Alias dereference policy (default: NeverDerefAliases).

data Filter

Conditions that must be fulfilled in order for the Search to match a given entry.

Constructors

Not !Filter
Filter does not match the entry

And !(NonEmpty Filter)
All filters match the entry

Or !(NonEmpty Filter)
Any filter matches the entry

Present !Attr
Attribute is present in the entry

!Attr := !AttrValue
Attribute's value is equal to the assertion

!Attr :>= !AttrValue
Attribute's value is equal to or greater than the assertion

!Attr :<= !AttrValue
Attribute's value is equal to or less than the assertion

!Attr :~= !AttrValue
Attribute's value approximately matches the assertion

!Attr :=* !(Maybe AttrValue, [AttrValue], Maybe AttrValue)
Glob match

(Maybe Attr, Maybe Attr, Bool) ::= AttrValue
Extensible match

Modify

modify :: Ldap -> Dn -> [Operation] -> IO ()

Perform the Modify operation synchronously. Raises ResponseError on failures.

data Operation

Type of modification being performed.

Constructors

Delete Attr [AttrValue]
Delete values from the attribute. Deletes the attribute if the list is empty or all current values are listed.

Add Attr [AttrValue]
Add values to the attribute, creating it if necessary.

Replace Attr [AttrValue]
Replace all existing values of the attribute with the new list. Deletes the attribute if the list is empty.

Add

add :: Ldap -> Dn -> AttrList NonEmpty -> IO ()

Perform the Add operation synchronously. Raises ResponseError on failures.

Delete

delete :: Ldap -> Dn -> IO ()

Perform the Delete operation synchronously. Raises ResponseError on failures.

ModifyDn

newtype RelativeDn

A component of Dn.

Constructors

RelativeDn Text

modifyDn :: Ldap -> Dn -> RelativeDn -> Bool -> Maybe Dn -> IO ()

Perform the Modify DN operation synchronously. Raises ResponseError on failures.

Compare

compare :: Ldap -> Dn -> Attr -> AttrValue -> IO Bool

Perform the Compare operation synchronously. Raises ResponseError on failures.

Extended

newtype Oid

Globally unique LDAP object identifier.

Constructors

Oid Text

extended :: Ldap -> Oid -> Maybe ByteString -> IO ()

Perform the Extended operation synchronously. Raises ResponseError on failures.

Waiting for completion

wait :: Async a -> IO (Either ResponseError a)

Wait for operation completion.

Miscellanous

newtype Dn

Unique identifier of an LDAP entry.

Constructors

Dn Text

newtype Attr

Attribute name.

Constructors

Attr Text

type AttrValue = ByteString

Attribute value.

type AttrList f = [(Attr, f AttrValue)]

List of attributes and their values. f is the structure these values are in, e.g. NonEmpty.

Re-exports

data NonEmpty a :: * -> *