Ldap.Client.Bind

Documentation

bind :: Ldap -> Dn -> Password -> IO ()

bindEither :: Ldap -> Dn -> Password -> IO (Either ResponseError ())

bindAsync :: Ldap -> Dn -> Password -> IO (Async ())

bindAsyncSTM :: Ldap -> Dn -> Password -> STM (Async ())

unbindAsync :: Ldap -> IO ()

Note that unbindAsync does not return an Async, because LDAP server never responds to UnbindRequests, hence a call to wait on a hypothetical Async would have resulted in an exception anyway.

unbindAsyncSTM :: Ldap -> STM ()

Note that unbindAsyncSTM does not return an Async, because LDAP server never responds to UnbindRequests, hence a call to wait on a hypothetical Async would have resulted in an exception anyway.