Ldap.Client.Compare

Description

Compare operation.

This operation comes in four flavours:

synchronous, exception throwing (compare)

synchronous, returning Either ResponseError () (compareEither)

asynchronous, IO based (compareAsync)

asynchronous, STM based (compareAsyncSTM)

Of those, the first one (compare) is probably the most useful for the typical usecase.

Documentation

compare :: Ldap -> Dn -> Attr -> AttrValue -> IO Bool

Perform the Compare operation synchronously. Raises ResponseError on failures.

compareEither :: Ldap -> Dn -> Attr -> AttrValue -> IO (Either ResponseError Bool)

Perform the Compare operation synchronously. Returns Left e where e is a ResponseError on failures.

compareAsync :: Ldap -> Dn -> Attr -> AttrValue -> IO (Async Bool)

Perform the Compare operation asynchronously. Call wait to wait for its completion.

compareAsyncSTM :: Ldap -> Dn -> Attr -> AttrValue -> STM (Async Bool)

Perform the Compare operation asynchronously.

Don't wait for its completion (with waitSTM) in the same transaction you've performed it in.

data Async a

Asynchronous LDAP operation. Use wait or waitSTM to wait for its completion.

wait :: Async a -> IO (Either ResponseError a)

Wait for operation completion.

waitSTM :: Async a -> STM (Either ResponseError a)

Wait for operation completion inside STM.

Do not use this inside the same STM transaction the operation was requested in! To give LDAP the chance to respond to it that transaction should commit. After that, applying waitSTM to the corresponding Async starts to make sense.