Ldap.Client.Extended

Description

Extended operation.

This operation comes in four flavours:

synchronous, exception throwing (extended)

synchronous, returning Either ResponseError () (extendedEither)

asynchronous, IO based (extendedAsync)

asynchronous, STM based (extendedAsyncSTM)

Of those, the first one (extended) is probably the most useful for the typical usecase.

Extended Operation

newtype Oid

Globally unique LDAP object identifier.

Constructors

Oid Text

extended :: Ldap -> Oid -> Maybe ByteString -> IO ()

Perform the Extended operation synchronously. Raises ResponseError on failures.

extendedEither :: Ldap -> Oid -> Maybe ByteString -> IO (Either ResponseError ())

Perform the Extended operation synchronously. Returns Left e where e is a ResponseError on failures.

extendedAsync :: Ldap -> Oid -> Maybe ByteString -> IO (Async ())

Perform the Extended operation asynchronously. Call wait to wait for its completion.

extendedAsyncSTM :: Ldap -> Oid -> Maybe ByteString -> STM (Async ())

Perform the Extended operation asynchronously.

Don't wait for its completion (with waitSTM) in the same transaction you've performed it in.

StartTLS Operation

startTls :: Ldap -> IO ()

An example of Extended Operation, cf. extended.

startTlsEither :: Ldap -> IO (Either ResponseError ())

An example of Extended Operation, cf. extendedEither.

startTlsAsync :: Ldap -> IO (Async ())

An example of Extended Operation, cf. extendedAsync.

startTlsAsyncSTM :: Ldap -> STM (Async ())

An example of Extended Operation, cf. extendedAsyncSTM.

data Async a

Asynchronous LDAP operation. Use wait or waitSTM to wait for its completion.

wait :: Async a -> IO (Either ResponseError a)

Wait for operation completion.

waitSTM :: Async a -> STM (Either ResponseError a)

Wait for operation completion inside STM.

Do not use this inside the same STM transaction the operation was requested in! To give LDAP the chance to respond to it that transaction should commit. After that, applying waitSTM to the corresponding Async starts to make sense.