Ldap.Client.Internal

Documentation

data Host

LDAP host.

Constructors

Plain String
Plain LDAP. Do not use!

Insecure String
LDAP over TLS without the certificate validity check. Only use for testing!

Secure String
LDAP over TLS. Use!

data PortNumber :: *

data Ldap

A token. All functions that interact with the Directory require one.

Constructors

Ldap

Fields

client :: TQueue ClientMessage

data ClientMessage

Constructors

New Request (TMVar (NonEmpty ProtocolServerOp))

data ResultCode

LDAP operation's result.

Constructors

Success

OperationError

ProtocolError

TimeLimitExceeded

SizeLimitExceeded

CompareFalse

CompareTrue

AuthMethodNotSupported

StrongerAuthRequired

Referral

AdminLimitExceeded

UnavailableCriticalExtension

ConfidentialityRequired

SaslBindInProgress

NoSuchAttribute

UndefinedAttributeType

InappropriateMatching

ConstraintViolation

AttributeOrValueExists

InvalidAttributeSyntax

NoSuchObject

AliasProblem

InvalidDNSyntax

AliasDereferencingProblem

InappropriateAuthentication

InvalidCredentials

InsufficientAccessRights

Busy

Unavailable

UnwillingToPerform

LoopDetect

NamingViolation

ObjectClassViolation

NotAllowedOnNonLeaf

NotAllowedOnRDN

EntryAlreadyExists

ObjectClassModsProhibited

AffectsMultipleDSAs

Other

data Async a

Asynchronous LDAP operation. Use wait or waitSTM to wait for its completion.

type AttrList f = [(Attr, f AttrValue)]

List of attributes and their values. f is the structure these values are in, e.g. NonEmpty.

Waiting for Request Completion

wait :: Async a -> IO (Either ResponseError a)

Wait for operation completion.

waitSTM :: Async a -> STM (Either ResponseError a)

Wait for operation completion inside STM.

Do not use this inside the same STM transaction the operation was requested in! To give LDAP the chance to respond to it that transaction should commit. After that, applying waitSTM to the corresponding Async starts to make sense.

Misc

type Response = NonEmpty InMessage

data ResponseError

Response indicates a failed operation.

Constructors

ResponseInvalid Request Response
LDAP server did not follow the protocol, so ldap-client couldn't make sense of the response.

ResponseErrorCode Request ResultCode Dn Text
The response contains a result code indicating failure and an error message.

type Request = ProtocolClientOp

raise :: Exception e => Either e a -> IO a

sendRequest :: Ldap -> (Response -> Either ResponseError a) -> Request -> STM (Async a)

newtype Dn

Unique identifier of an LDAP entry.

Constructors

Dn Text

newtype Attr

Attribute name.

Constructors

Attr Text

type AttrValue = ByteString

Attribute value.

unAttr :: Attr -> Text

Unbind operation

unbindAsync :: Ldap -> IO ()

Terminate the connection to the Directory.

Note that unbindAsync does not return an Async, because LDAP server never responds to UnbindRequests, hence a call to wait on a hypothetical Async would have resulted in an exception anyway.

unbindAsyncSTM :: Ldap -> STM ()

Terminate the connection to the Directory.

Note that unbindAsyncSTM does not return an Async, because LDAP server never responds to UnbindRequests, hence a call to wait on a hypothetical Async would have resulted in an exception anyway.