Ldap.Client.Modify

Description

Modify and Modify DN operations.

These operations come in four flavours:

synchronous, exception throwing (modify / modifyDn)

synchronous, returning Either ResponseError () (modifyEither / modifyDnEither)

asynchronous, IO based (modifyAsync / modifyDnAsync)

asynchronous, STM based (modifyAsyncSTM / modifyDnAsyncSTM)

Of those, the first one (modify / modifyDn) is probably the most useful for the typical usecase.

Documentation

data Operation

Type of modification being performed.

Constructors

Delete Attr [AttrValue]
Delete values from the attribute. Deletes the attribute if the list is empty or all current values are listed.

Add Attr [AttrValue]
Add values to the attribute, creating it if necessary.

Replace Attr [AttrValue]
Replace all existing values of the attribute with the new list. Deletes the attribute if the list is empty.

modify :: Ldap -> Dn -> [Operation] -> IO ()

Perform the Modify operation synchronously. Raises ResponseError on failures.

modifyEither :: Ldap -> Dn -> [Operation] -> IO (Either ResponseError ())

Perform the Modify operation synchronously. Returns Left e where e is a ResponseError on failures.

modifyAsync :: Ldap -> Dn -> [Operation] -> IO (Async ())

Perform the Modify operation asynchronously. Call wait to wait for its completion.

modifyAsyncSTM :: Ldap -> Dn -> [Operation] -> STM (Async ())

Perform the Modify operation asynchronously.

Don't wait for its completion (with waitSTM) in the same transaction you've performed it in.

newtype RelativeDn

A component of Dn.

Constructors

RelativeDn Text

modifyDn :: Ldap -> Dn -> RelativeDn -> Bool -> Maybe Dn -> IO ()

Perform the Modify DN operation synchronously. Raises ResponseError on failures.

modifyDnEither :: Ldap -> Dn -> RelativeDn -> Bool -> Maybe Dn -> IO (Either ResponseError ())

Perform the Modify DN operation synchronously. Returns Left e where e is a ResponseError on failures.

modifyDnAsync :: Ldap -> Dn -> RelativeDn -> Bool -> Maybe Dn -> IO (Async ())

Perform the Modify DN operation asynchronously. Call wait to wait for its completion.

modifyDnAsyncSTM :: Ldap -> Dn -> RelativeDn -> Bool -> Maybe Dn -> STM (Async ())

Perform the Modify DN operation asynchronously.

Don't wait for its completion (with waitSTM) in the same transaction you've performed it in.